Added a new conditional test IP? which tests if the number in X is an integer or not.  Skipping if not.  This command is in the test catalogue.  It works in real and integer modes (although it is always true in the latter case).
Added te converse conditional test FP? which tests if the number
in X has a fractional component or not.  Skipping if not.  Again in
the test catalogue and it works in real and integer modes (always testing
false in the latter case).

Added a new command SKIP which takes an argument 0 - 99 and skips
that number of program steps forwards.  If an attempt is made to
skip past the end of the program, it is treated as if the program
ran off the end and does an implicit return.  This command is in
the programming catalogue and can take an indirect argument as usual.

Added a new command BACK which acts like SKIP but jumps backwards instead
of forwards.  Going back before the start of the program (step 0 or before)
stops program execution.  This command is also in the programming catalogue.


Cleaned up some of the code.  Made more little wrapper functions.  Thumb isn't
great at bit field manipulations and a function call is usually shorter.